+++ mcp_chat/components/__init__.py
