+++ mcp_chat/components/loading_icon.py
