+++ mcp_chat/containers.py
@@ -22,7 +22,7 @@ from langgraph.store.memory import InMemoryStore
 # from langgraph.checkpoint.postgres.aio import AsyncPostgresSaver, AsyncShallowPostgresSaver
 # from langgraph.checkpoint.sqlite.aio import AsyncSqliteSaver
 # from langgraph.store.postgres import AsyncPostgresStore
-from host_app.mcp_client import MultiMCPClient, SSEConnection, StdioConnection
+from mcp_chat.mcp_client import MultiMCPClient, SSEConnection, StdioConnection
 
 # Load .env file into environment variables (so they can be used in config.yml)
 load_dotenv()
