+++ rxconfig.py
@@ -1,5 +1,5 @@
 import reflex as rx
 
 config = rx.Config(
-    app_name="host_app",
+    app_name="mcp_chat",
 )
